Server: return proper messages and status codes in notes controller

+++ backend/src/controllers/note.js
@@ -5,14 +5,34 @@ const UserModel = require('../models/User');
 
 module.exports = {
   async index(req, res) {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res
+        .status(422)
+        .json({ message: 'Invalid input', errors: errors.array() });
+    }
+
+    if (!req.isAuthenticated) {
+      return res.status(401).json({ message: 'User is not authenticated' });
+    }
+
+    const { userId } = req.params;
+
+    if (req.userId !== userId) {
+      return res
+        .status(401)
+        .json({ message: 'User is not authorized to perform this action' });
+    }
+
+    const { isValid, errorMessage } = await checkTokenValidation(
+      userId,
+      req.token
+    );
+    if (!isValid) {
+      return res.status(401).json({ message: errorMessage });
+    }
+
     try {
-      // Finds the validation errors in this request and wraps them in an object with handy functions
-      const errors = validationResult(req);
-      if (!errors.isEmpty())
-        return res.status(422).json({ errors: errors.array() });
-
-      const { userId } = req.params;
-
       const user = await UserModel.findById(userId);
       if (!user) throw new Error('User does not exists');
 
@@ -25,9 +45,31 @@ module.exports = {
   },
 
   async show(req, res) {
+    if (!req.isAuthenticated) {
+      return res.status(401).json({ message: 'User is not authenticated' });
+    }
+
+    const { noteId } = req.params;
+
     try {
-      const { noteId } = req.params;
       const note = await NoteModel.findById(noteId);
+
+      const userId = note.creator._id;
+
+      if (req.userId !== userId) {
+        return res
+          .status(401)
+          .json({ message: 'User is not authorized to perform this action' });
+      }
+
+      const { isValid, errorMessage } = await checkTokenValidation(
+        userId,
+        req.token
+      );
+      if (!isValid) {
+        return res.status(401).json({ message: errorMessage });
+      }
+
       return res.status(200).json({ note });
     } catch (error) {
       return res.status(500).json({ message: error.message });
@@ -35,15 +77,36 @@ module.exports = {
   },
 
   async store(req, res) {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res
+        .status(422)
+        .json({ message: 'Invalid input', errors: errors.array() });
+    }
+
+    if (!req.isAuthenticated) {
+      return res.status(401).json({ message: 'User is not authenticated' });
+    }
+
+    const { userId } = req.params;
+
+    if (req.userId !== userId) {
+      return res
+        .status(401)
+        .json({ message: 'User is not authorized to perform this action' });
+    }
+
+    const { isValid, errorMessage } = await checkTokenValidation(
+      userId,
+      req.token
+    );
+    if (!isValid) {
+      return res.status(401).json({ message: errorMessage });
+    }
+
+    const { title, description } = req.body;
+
     try {
-      // Finds the validation errors in this request and wraps them in an object with handy functions
-      const errors = validationResult(req);
-      if (!errors.isEmpty())
-        return res.status(422).json({ errors: errors.array() });
-
-      const { userId } = req.params;
-      const { title, description } = req.body;
-
       const user = await UserModel.findById(userId);
       if (!user) throw new Error('User does not exists');
 
@@ -53,38 +116,89 @@ module.exports = {
         creator: user
       });
 
-      return res.status(200).json({ createdNote });
+      return res.status(201).json({ note: createdNote });
     } catch (error) {
       return res.status(500).json({ message: error.message });
     }
   },
 
   async delete(req, res) {
+    if (!req.isAuthenticated) {
+      return res.status(401).json({ message: 'User is not authenticated' });
+    }
+
+    const { noteId } = req.params;
+
     try {
-      const { noteId } = req.params;
+      const note = await NoteModel.findById(noteId);
+      console.log('note ', note);
+      const userId = note.creator._id;
+      console.log('user ', userId);
+
+      if (req.userId !== userId.toString()) {
+        return res
+          .status(401)
+          .json({ message: 'User is not authorized to perform this action' });
+      }
+
+      const { isValid, errorMessage } = await checkTokenValidation(
+        userId.toString(),
+        req.token
+      );
+      console.log('is valid ', isValid);
+      if (!isValid) {
+        console.log('message ', errorMessage);
+        return res.status(401).json({ message: errorMessage });
+      }
+
       await NoteModel.findByIdAndDelete(noteId);
-      return res.status(204).end();
     } catch (error) {
       return res.status(500).json({ message: error.message });
     }
+
+    return res.status(204).end();
   },
 
   async update(req, res) {
-    try {
-      // Finds the validation errors in this request and wraps them in an object with handy functions
-      const errors = validationResult(req);
-      if (!errors.isEmpty())
-        return res.status(422).json({ errors: errors.array() });
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res
+        .status(422)
+        .json({ message: 'Invalid input', errors: errors.array() });
+    }
 
-      const { noteId } = req.params;
+    if (!req.isAuthenticated) {
+      return res.status(401).json({ message: 'User is not authenticated' });
+    }
+
+    const { noteId } = req.params;
+
+    try {
+      const note = await NoteModel.findById(noteId);
+      const userId = note.creator._id;
+
+      if (req.userId !== userId) {
+        return res
+          .status(401)
+          .json({ message: 'User is not authorized to perform this action' });
+      }
+
+      const { isValid, errorMessage } = await checkTokenValidation(
+        userId,
+        req.token
+      );
+      if (!isValid) {
+        return res.status(401).json({ message: errorMessage });
+      }
+
       const { title, description } = req.body;
 
       await NoteModel.findByIdAndUpdate(noteId, { title, description });
-
-      return res.status(204).end();
     } catch (error) {
       return res.status(500).json({ message: error.message });
     }
+
+    return res.status(204).end();
   },
 
   validate: function validate(method) {
